refactor(taxes): extract TaxReturnSection component
Move TaxReturnSection to src/pages/taxes/components/TaxReturnSection.tsx.
Remove unused imports (useRef, TaxChecklistItem, OwnerBadge) from orchestrator.
Taxes.tsx 548→413L.

Part of #181
